BAH-4858 | Fix. widgets build failing due to extracted fucntion export path
