+++ packages/bahmni-widgets/src/observations/index.ts
@@ -1,4 +1,4 @@
 export { default as Observations } from './Observations';
 export type { ObservationConfig } from './Observations';
-export { extractObservationsFromBundle } from './utils';
+export { extractObservationsFromBundle, extractFormName } from './utils';
 export type { ExtractedObservationsResult } from './models';
